perf: remove code that triggers indexing of included files (#149)
This code finds all `include` directives, then compares this list with
`beanFiles`. For all matches in triggers async indexing. But all files
from `beanFiles` are already scheduled for indexing in `onInitialized`.

With this code every time a file changed, all included files are also
indexed, even though they are not changed. And if they are, regular
indexing on change event would be triggered.